unify IncrementalDecoder and BaseDecoder (#586)
